requirements.txt
Flask==3.0.0
google-generativeai==0.3.2
notion-client==2.2.1
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.26.2
PyPDF2==3.0.1
python-docx==1.1.0
reportlab==4.0.7
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
